misc: massive cleanup
