--- chactivityserver/src/TODO
@@ -1,36 +1,0 @@
-+ Support follow/accept requests:
-
-{
-  "@context": "https://www.w3.org/ns/activitystreams",
-  "type": "Follow",
-  "id": "https://example.com/follow-request/12345",
-  "actor": "https://example.com/user/12345",
-  "object": "https://example.org/user/67890"
-}
-
-
-{
-  "@context": "https://www.w3.org/ns/activitystreams",
-  "type": "Accept",
-  "id": "https://example.com/accept/12345",
-  "actor": "https://example.com/server/12345",
-  "object": "https://example.com/follow-request/67890"
-}
-
-
-+ Générer outbox
-  + TODO group per 20/page
-  + TODO cache
-  + TODO date
-  + TODO validate
-
-+ Parse inbox from another instance
-
-+ Update crates to latest version
-
-+ TODO webfinger
-
-curl https://cha-cu.it/.well-known/webfinger\?resource\=acct:chef@cha-cu.it
-
-curl -H "Accept: application/activity+json" https://cha-cu.it/users/chef
-
