+++ chactivityserver/src/inbox.rs
@@ -1,0 +1,53 @@
+
+
+/*
+
+
+// Define the regular expression for matching the required header
+// TODO order or generate all?
+const REQUIRED_HEADER: &str = r#"^---\ntitle: ([^\n]+)\n?duration: ([^\n]+)\n?tags: (\[[^\]]+\])\n?thumbnail: ("[^"]+")\n?---\n"#;
+// Define the regular expression for matching the required titles
+const REQUIRED_TITLES: &str = r"# Ingrédients[\s\S]*# Équipement[\s\S]*# Préparation[\s\S]*# Notes";
+
+
+async fn inbox(data: web::Json<InboxRequest>) -> &'static str {
+    if data.activity_type == "Create" && data.object.object_type == "Article" && data.object.media_type == "text/markdown" {
+        // Compile the regular expression
+        let re_header = Regex::new(REQUIRED_HEADER).unwrap();
+        let re_titles = Regex::new(REQUIRED_TITLES).unwrap();
+
+        // Check if the content of the article matches the required titles
+        if re_header.is_match(&data.object.content) && re_titles.is_match(&data.object.content) {
+            // Get the current date
+            let now: DateTime<Local> = Local::now();
+            let current_date = now.format("%Y-%m-%d").to_string();
+
+            // Add the current date to the header of the article
+            let new_content = re_header.replace(
+                                                &data.object.content,
+                                                format!("---\ntitle: {}\nduration: $2\ntags: $3\nthumbnail: $4\ndate: {}\n---\n", data.object.name, current_date)
+                                            ).to_string();
+
+            // Save the content of the article to a file
+            let file_path = format!("chadiverse/{}.md", data.object.name);
+            // Create the "chadiverse" directory if it does not exist
+            if let Err(_) = fs::create_dir_all("chadiverse") {
+                return "Failed to create directory";
+            }
+
+            // Write the file
+            if let Err(_) = fs::write(file_path, new_content) {
+                return "Failed to write file";
+            }
+
+            // TODO add tests
+            println!("Received article '{}' from {}", data.object.name, data.actor);
+            "Article received"
+        } else {
+            "Article does not have the required titles"
+        }
+    } else {
+        "Activity received"
+    }
+}
+*/
